# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USC App Degree Projects API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Status"
            }
          },
          {
            "name": "modality_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality Id"
            }
          },
          {
            "name": "academic_period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Period"
            }
          },
          {
            "name": "academic_program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Program Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Inscribe una nueva idea de trabajo de grado (solo Estudiante).",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List Messages",
        "description": "Retorna todos los mensajes del trabajo de grado, ordenados por sent_at DESC.\nRequiere pertenencia activa (Admin, Director, Jurado o Estudiante del proyecto).\nEl campo sender_display ya incorpora el anonimato del jurado desde el momento\nen que se guardó el mensaje.",
        "operationId": "list_messages_api_v1_projects__project_id__messages_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Messages Api V1 Projects  Project Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send Message",
        "description": "Envía un mensaje a un actor del trabajo o a todos (solo Administrador puede\nomitir recipient_id para broadcast).\n\nReglas de mensajería (RF-15-02..RF-15-07):\n  Estudiante  → Director o Jurado\n  Director    → Estudiante o Administrador\n  Jurado      → Estudiante\n  Admin       → cualquiera (broadcast si recipient_id es null)\n\nEl anonimato del jurado se aplica en sender_display: si el emisor es Jurado,\nel campo se establece como \"Jurado N\" independientemente del receptor.",
        "operationId": "send_message_api_v1_projects__project_id__messages_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/messages/{msg_id}/read": {
      "patch": {
        "tags": [
          "messages"
        ],
        "summary": "Mark Message Read",
        "description": "Marca un mensaje como leído. Solo el receptor puede hacerlo.\nSi el mensaje ya está leído, responde 200 (idempotente).",
        "operationId": "mark_message_read_api_v1_projects__project_id__messages__msg_id__read_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "msg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Msg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/unread-count": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Get Unread Count",
        "description": "Retorna el número de mensajes no leídos dirigidos al usuario autenticado.\nUsa un COUNT directo para eficiencia — no trae filas completas.",
        "operationId": "get_unread_count_api_v1_messages_unread_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "recipient_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Id"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MessageCreate",
        "example": {
          "content": "Por favor revisar el capítulo 3 antes del viernes.",
          "recipient_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        }
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sender_display": {
            "type": "string",
            "title": "Sender Display"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sender_display",
          "content",
          "is_read",
          "sent_at"
        ],
        "title": "MessageResponse",
        "example": {
          "content": "Por favor revisar el capítulo 3 antes del viernes.",
          "id": "f0a1b2c3-0abc-1234-5678-9abcdef01234",
          "is_read": false,
          "sender_display": "Dr. Carlos Rodríguez",
          "sent_at": "2026-03-10T14:30:00Z"
        }
      },
...
      "PaginatedProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedProjectsResponse",
        "example": {
          "items": [],
          "page": 1,
          "size": 20,
          "total": 0
        }
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "maxLength": 200,
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "member_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Member Ids"
          },
          "prerequisite_declaration": {
            "type": "boolean",
            "title": "Prerequisite Declaration"
          }
        },
        "type": "object",
        "required": [
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "member_ids",
          "prerequisite_declaration"
        ],
        "title": "ProjectCreate",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "member_ids": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          ],
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "prerequisite_declaration": true,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC"
        }
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "has_company_link": {
            "type": "boolean",
            "title": "Has Company Link"
          },
          "plagiarism_suspended": {
            "type": "boolean",
            "title": "Plagiarism Suspended"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "suggested_director",
          "period",
          "status",
          "has_company_link",
          "plagiarism_suspended",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "created_at": "2026-02-05T09:00:00Z",
          "has_company_link": false,
          "id": "f6a7b8c9-6789-abcd-f012-345678901234",
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "period": "2026-1",
          "plagiarism_suspended": false,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "status": "idea_inscrita",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC",
          "updated_at": "2026-02-05T09:00:00Z"
        }
      },
...
      "UnreadCountResponse": {
        "properties": {
          "unread": {
            "type": "integer",
            "title": "Unread"
          }
        },
        "type": "object",
        "required": [
          "unread"
        ],
        "title": "UnreadCountResponse",
        "example": {
          "unread": 3
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}